+++ scripts/README.md
@@ -1,0 +1,1 @@
+These scripts are used to interpolate data to roms grid
