+++ scripts/clear
@@ -1,0 +1,2 @@
+#!/bin/bash
+rm *nc
