--- scripts/era_glorys/mask_change.txt
@@ -1,27 +1,0 @@
-153 143 1.000000
-157 157 1.000000
-161 153 1.000000
-155 142 1.000000
-161 153 0.000000
-155 142 0.000000
-155 143 1.000000
-156 160 1.000000
-156 159 1.000000
-136 156 1.000000
-136 156 0.000000
-138 156 1.000000
-139 157 1.000000
-137 156 1.000000
-138 156 0.000000
-137 156 0.000000
-139 156 1.000000
-139 156 0.000000
-139 156 1.000000
-140 157 1.000000
-154 156 1.000000
-156 159 0.000000
-156 160 0.000000
-154 156 0.000000
-140 157 0.000000
-139 157 0.000000
-139 156 0.000000
